GEODE-7944: Fixes type mistmatch
* size_t can't implicitly cast to int32_t on MSVC.

+++ cppcache/src/TcrConnection.cpp
@@ -164,7 +164,7 @@ bool TcrConnection::initTcrConnection(
     const auto& dsMemberId = memId->getDSMemberId();
     handShakeMsg.writeBytes(
         reinterpret_cast<const uint8_t*>(dsMemberId.c_str()),
-        dsMemberId.size());
+        static_cast<int32_t>(dsMemberId.size()));
   } else {
     ACE_TCHAR hostName[256];
     ACE_OS::hostname(hostName, sizeof(hostName) - 1);
@@ -184,7 +184,7 @@ bool TcrConnection::initTcrConnection(
     const auto& dsMemberId = memId->getDSMemberId();
     handShakeMsg.writeBytes(
         reinterpret_cast<const uint8_t*>(dsMemberId.c_str()),
-        dsMemberId.size());
+        static_cast<int32_t>(dsMemberId.size()));
   }
   handShakeMsg.writeInt(static_cast<int32_t>(1));
